vote buttons exist

+++ app/controllers/boards_controller.rb
@@ -1,19 +1,19 @@
 class BoardsController < ApplicationController
-
+  # before_action except: [:show, :edit, :update, :destroy]
   skip_before_action :authenticate_user!, only: [:index]
 
   def new
     @board = Board.new
+    authorize @board
   end
 
   def index
     @boards = Board.all
-     # authorize @boards
   end
 
   def create
     @board = Board.new(topic: params[:board][:topic])
-    # authorize @board
+    authorize @board
     if @board.save
       flash[:notice] = "Board created!"
       redirect_to @board
@@ -24,13 +24,20 @@ class BoardsController < ApplicationController
 
   def show
     @board = Board.find params[:id]
-    # authorize @board
+    authorize @board
   end
 
   def destroy
     board = current_user.boards.find params[:id]
-    # authorize board
+    authorize board
     board.destroy
     redirect_to "/boards"
   end
+
+  private
+
+  # def set_board
+  #   @board = Board.find(params[:board_id])
+  # end
+
 end
